03-project-management-app/pyproject.toml
[tool.poetry]
name = "03-project-management-app"
version = "0.1.0"
description = "Creating a FastAPI backend to delve into SQLAlchemy and Alembic, mastering ORM, migrations, and schema management for efficient database operations. The project structure mirrors a real-world project management system, with features like project creation, task assignment, progress tracking, and reporting."
authors = ["Rashmin Dungrani <rashmindungrani@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
orjson = "^3.10.3"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.1.3"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
